requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
pytest
pytest-flask
coverage
pytest-cov